+++ txtgengui/Builds/MacOSX/CHANGES.md
@@ -1,3 +1,11 @@
+#  version 1.2.0
+
+## program mode
+can be programmed in lua
+with simple instructions to speak and show text
+with grammar and LSystem
+onKey
+
 #  version 1.0.1
 
 ## markov